% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjR2.R
\name{adjR2}
\alias{adjR2}
\title{Adjusted R2}
\usage{
adjR2(pred, truth, d)
}
\arguments{
\item{pred}{a numeric vector of predicted/fitted values}

\item{truth}{a numeric vector of true/observed values}

\item{d}{number of prediction variables used}
}
\value{
adjusted r2-value
}
\description{
Computes the adjusted R2 value
}
\details{
Uses 2 numeric vectors pred and truth to compute an adjusted R2 value, adjusted by number of parameters
}
\examples{
#using the ISLR::Credit database for this example
truth <- ISLR::Credit$Limit
reg_model <- lm(ISLR::Credit$Limit~ISLR::Credit$Income + ISLR::Credit$Cards)
pred <- reg_model$fitted.values
adjR2(pred, truth, 2)

}
