% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxMuller.R
\name{boxMuller}
\alias{boxMuller}
\title{Computes Box Muller Transformation}
\usage{
boxMuller(n)
}
\arguments{
\item{n}{number of sample estimates from the normal distribution}
}
\value{
a vector of doubles of size n sampled from the normal standard distribution N(0,1)
}
\description{
takes a desired value for n and returns n samples from the standard normal distribution using Box Muller Transformation
}
\details{
computes n samples using the Box Muller transformation to estimate samples from a normal distribution. Uses the following two formulas: \cr
\eqn{x1 = sqrt(-2*log(u1)) * (cos(2*pi*u2))} \cr
\eqn{x2 = sqrt(-2*log(u1)) * (sin(2*pi*u2))}
}
\examples{
normal_estimate <- boxMuller(500)
#verify that mu and std is approximately 0 and 1
mean(normal_estimate)
sd(normal_estimate)
}
