% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSquareCounts.R
\name{chiSquareCounts}
\alias{chiSquareCounts}
\title{Chi-Squared Counts}
\usage{
chiSquareCounts(tib, v1, v2)
}
\arguments{
\item{tib}{a tibble of data with at least 2 columns represting grouping variables}

\item{v1}{string representing column name of first grouping variable in the tibble}

\item{v2}{string representing column name of second grouping variable in the tibble}
}
\value{
Chi Sq test statistic object, with X2, df, and p-values
}
\description{
Given a tibble, calculates a chi-square statistic, with degree of freedom and p-values
}
\details{
computes a 2x2 count table, uses chisq.test() to compute the chi-sq statistics and results
}
\examples{
#using a sample daaset from Dr. McKay's website
count_data <- read.csv("https://jlucasmckay.bmi.emory.edu/global/bmi585/demographics.csv")
chiSquareCounts(count_data,"sex","group")
}
