% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSize.R
\name{effectSize}
\alias{effectSize}
\title{Effect Size}
\usage{
effectSize(x, g)
}
\arguments{
\item{x}{a vector of doubles with the observations}

\item{g}{a vector of 1 or 0 indicating grouping.}
}
\value{
Effect Size / Cohen's D
}
\description{
Calculates the effect size / Cohen's D given a vector of data and a vector for grouping
}
\details{
Computes the effect size by dividing the sample difference by the standard deviation of the pooled sample.
The function takes a vector of observations and a vector of groups in the form of group1 == 1, group2 ==2.
Group 1 and 2 must of same size
}
\examples{

groups <- c(rep(1,times=50),rep(0,times=50))  \cr
modeled_tumor_size <- 4 -2*groups + rnorm(100)  \cr
effectSize(modeled_tumor_size, groups) \cr
}
