% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdrAdjust.R
\name{fdrAdjust}
\alias{fdrAdjust}
\title{FDR (False Discovery Rate) Adjustment}
\usage{
fdrAdjust(p)
}
\arguments{
\item{p}{a numeric vector of p values}
}
\value{
logical vector, true if p values remain significant after bonferroni correction
}
\description{
Tests for significance using fdr adjustment for multiple hypothesis testing
}
\details{
Tests to see if p-values remain significant at p < 0.05 using fdr adjustment for multiple-hypothesis testing
}
\examples{
p <- c(0.0025, 0.0050, 0.0075, 0.0100, 0.0125, 0.0150, 0.0175, 0.0200, 0.0225, 0.0250)
fdrAdjust(p)
}
