% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{getCounts}
\alias{getCounts}
\title{Get Counts for 2x2 Table}
\usage{
getCounts(pred, truth)
}
\arguments{
\item{pred}{vector of 1/0 or True and False representing predictions}

\item{truth}{vector of 1/0 or True and False representing gold standards or ground truths}
}
\value{
the sensitivity given as true pos/(true pos + false neg)
}
\description{
Support function that takes a set of predictions and set of ground truths
predictions and ground truths can either be boolean or 1 and 0, must be same length
Returns numbers of True Positive, True Negatives, False Positives, and False Negatives in that order in a number vector of size 4
}
