% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiPaul.R
\name{hiPaul}
\alias{hiPaul}
\title{Hi Paul}
\usage{
hiPaul()
}
\description{
this is a function that just says hi Paul
}
\details{
returns the string "hello world, hello Paul. tHis is the most world shattering function ever
}
\examples{
hiPaul()

}
