% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Minimum N}
\usage{
minimumN(d)
}
\arguments{
\item{d}{a desired effect size / Cohen's D}
}
\value{
minimum number of samples needed to resolve effect size d rounded up to nearest whole number
}
\description{
Computes the minimum number of samples n per group to resolve a specified effect size d with 80\% power
}
\details{
uses power.t.test to compute minimum sample sizes n. Passes 0.8 to power.t.test to compute sample size for 80\% power
Assumes discrete number of observations, and so the ceiling of the output is returned.
}
\examples{
minimumN(0.5)
}
