% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{PC Approximation}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{a tibble or dataframe}

\item{npc}{number of principal components to use in estimate}
}
\description{
returns an approximation to the data based on a number of principal components
}
\details{
takes a tibble or dataframe, a number of principal components to include, and returns an n x m dimensional estimate such that n is the number of observations and m = npc
}
\examples{
X <- tibble(iris[,-which(names(iris)=="Species")]) #iris dataset without species
pcApprox(X, 2)
}
