% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcLollipop.R
\name{pcLollipop}
\alias{pcLollipop}
\title{Principal Component Lollipop Plot}
\usage{
pcLollipop(x)
}
\arguments{
\item{x}{a tibble or dataframe}

\item{npc}{number of principal components to use in estimate}
}
\description{
plots the lollipop plot from principal component analysis given a tibble
}
\details{
takes a tibble or dataframe and plots the loading matrix computed by principal component analysis. Uses ggplot to render graphic
}
\examples{
x <- iris[1:4]
pcLollipop(x)
}
