% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postHocPower.R
\name{postHocPower}
\alias{postHocPower}
\title{Post-Hoc Power}
\usage{
postHocPower(d, n1, n2)
}
\arguments{
\item{d}{effect size}

\item{n1}{number of observations in first group}

\item{n2}{number of observations in second group}
}
\value{
post hoc power estimate
}
\description{
Given an effect size d and two sample sizes n1 and n2, estimates Post-Hoc Power
}
\details{
Uses Monte Carlo simulation approach to run 1000 t-tests against 2 generated samples of size n1 and n2 respectively.
The mean of the second normal distribution is centered at d to allow for the given effect size
}
\examples{
postHocPower(1,12,12)

}
