% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{R2}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{a numeric vector of predicted/fitted values}

\item{truth}{a numeric vector of true/observed values}
}
\value{
r2-value
}
\description{
Computes the R2 value
}
\details{
Uses 2 numeric vectors pred and truth to compute an R2 value
}
\examples{
#using the ISLR::Credit database for this example
truth <- ISLR::Credit$Limit
reg_model <- lm(ISLR::Credit$Limit~ISLR::Credit$Income)
pred <- reg_model$fitted.values
r2(pred, truth)
}
