% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoSidedT.R
\name{twoSidedT}
\alias{twoSidedT}
\title{Two Sided T-Test}
\usage{
twoSidedT(t, df)
}
\arguments{
\item{t}{a t-statistic to compute area}

\item{df}{degrees of freedom}
}
\value{
two sided area under t-density given t, df
}
\description{
Takes a test statistic t, a degree of freedom and computes the two sided area under the t-density
}
\details{
Computes two sided area under t-density using R's pt function. Left sided area under density is computed, and doubled.
}
\examples{
twoSidedT(2, df=10)
}
