% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoSidedZ.R
\name{twoSidedZ}
\alias{twoSidedZ}
\title{Two Sided Z-Score}
\usage{
twoSidedZ(z)
}
\arguments{
\item{z}{a value for the z-statistic}
}
\value{
two sided area under z-density given a value for z statistic
}
\description{
Take a Z-statistic and computes two sided area under the Z-desnity
}
\details{
Computes two sided area under z-density using R's pnorm function. Left sided area under density is computed, and doubled.
}
\examples{
twoSidedZ(2)
}
