% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale}
\usage{
unscale(x)
}
\arguments{
\item{x}{a numerical function that has been scaled (with scale())}
}
\value{
unscaled values for x
}
\description{
Given a scaled numerical object, unscales back to original
}
\details{
Uses attributes of a scaled object to return it back to the original
}
\examples{
x <- matrix(1:10)
scaled_x <- scale(x)
unscale(scaled_x)

x2 <- matrix(1:10, ncol=5)
scaled_x2 <- scale(x2)
unscale(scaled_x2)
}
