% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welchT.R
\name{welchT}
\alias{welchT}
\title{Welch T-Test}
\usage{
welchT(x, y)
}
\arguments{
\item{x}{numeric vector representing sample 1}

\item{y}{numeric vector representing sample 2}
}
\value{
2 named variables: t-statistic, effective degree of freedom
}
\description{
Computes a manual Welch T-test given two sample vectors
}
\details{
Takes two numeric vectors x, y representing two samples.
Computes a manual Welch T-test statistic and effective degrees of freedom
}
\examples{
#using an example from class with data by Dr. McKay:
bp_df <- read.csv("https://jlucasmckay.bmi.emory.edu/global/bmi585/two_group_bp.csv")
group_a <- bp_df$bp[which(bp_df$group=='A')]
group_b <- bp_df$bp[which(bp_df$group=='B')]
welchT(group_a,group_b)
}
