% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Calculates Accuracy}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{vector of 1/0 or True and False representing predictions}

\item{truth}{vector of 1/0 or True and False representing gold standards or ground truths}
}
\value{
Accuracy, computed as (true pos + true neg) / (all pos + all neg)
}
\description{
Calculates the accuracy of a given test/prediction against a gold standard/ground truth.
}
\details{
Uses getCounts to obtain values for the 2x2 table
predictions and ground truths can either be boolean or 1 and 0, must be same length
}
