% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxMuller.R
\name{boxMuller}
\alias{boxMuller}
\title{Calculates Box Muller}
\usage{
boxMuller(n)
}
\arguments{
\item{n}{number of observations}
}
\value{
stuff
}
\description{
takes an N and does box muller
}
\details{
to write later
}
