% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv.R
\name{ppv}
\alias{ppv}
\title{Calculates Positive Predictive Value}
\usage{
ppv(pred, truth)
}
\arguments{
\item{pred}{vector of 1/0 or True and False representing predictions}

\item{truth}{vector of 1/0 or True and False representing gold standards or ground truths}
}
\value{
Positive predicitive value (PPV)
}
\description{
Calculates the Positive Predictive Value (PPV) of a given test/prediction against a gold standard/ground truth.
}
\details{
Calculates the specificity of a given test/prediction against a gold standard/ground truth. Predictions and ground truths can either be boolean or 1 and 0, must be same length
}
