% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivities.R
\name{sensitivity}
\alias{sensitivity}
\title{Computers Sensitivities}
\usage{
sensitivity(pred, truth)
}
\arguments{
\item{pred}{vector of 1/0 or True and False representing predictions}

\item{truth}{vector of 1/0 or True and False representing gold standards or ground truths}
}
\value{
the sensitivity given as true pos/(true pos + false neg)
}
\description{
gets sensitivity given set of predictions and set of ground truths
predictions and ground truths can either be boolean or 1 and 0, must be same length
}
\examples{
sensitivity(c(T,T,F,T), c(T,T,F,F))
}
