% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{specificity}
\alias{specificity}
\title{Calculates Specificity}
\usage{
specificity(pred, truth)
}
\arguments{
\item{pred}{vector of 1/0 or True and False representing predictions}

\item{truth}{vector of 1/0 or True and False representing gold standards or ground truths}
}
\value{
Specificity, computed as true neg / (true neg + false pos)
}
\description{
Calculates the specificity of a given test/prediction against a gold standard/ground truth.
}
\details{
Calculates the specificity of a given test/prediction against a gold standard/ground truth. Predictions and ground truths can either be boolean or 1 and 0, must be same length
}
